GUVNOR-2841 : Guided Decision Table: Restored dtable should not be read only
